+++ Entities/Review.cs
@@ -2,8 +2,8 @@ namespace VirtualLibraryAPI.Entities
 {
     public class Review
     {
-        public int      Id        { get; set; }
-        public int      BookId    { get; set; }
+        public long      Id        { get; set; }
+        public long      BookId    { get; set; }
         public Book     Book      { get; set; } = null!;
         public long      UserId    { get; set; }
         public string   UserName  { get; set; } = null!;   // denormalize for display
